Move db import and export contracts to main activity in order to keep the utils package as clean as possible
