--- app/src/main/java/app/musikus/utils/dbExportImport.kt
@@ -1,33 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * Copyright (c) 2022 Matthias Emde
- */
-
-package app.musikus.utils
-
-import android.content.Context
-import android.content.Intent
-import android.os.Build
-import android.os.Environment
-import android.provider.DocumentsContract
-import androidx.activity.result.contract.ActivityResultContracts
-
-class ExportDatabaseContract : ActivityResultContracts.CreateDocument("*/*") {
-    override fun createIntent(context: Context, input: String) =
-        super.createIntent(context, input).apply {
-            addCategory(Intent.CATEGORY_OPENABLE)
-            if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                putExtra(DocumentsContract.EXTRA_INITIAL_URI, Environment.DIRECTORY_DOWNLOADS)
-            }
-        }
-}
-
-class ImportDatabaseContract : ActivityResultContracts.OpenDocument() {
-    override fun createIntent(context: Context, input: Array<String>) =
-        super.createIntent(context, input).apply {
-
-        }
-}
